DVBAPI: Improve reopenonzap functionality to check for otherwise used ca device (recording) and so that device is not opened/closed multiple times on zap anymore. Changed some integers to explicit size definitions.
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@6359 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-cccshare.c
@@ -1319,7 +1319,7 @@ void done_share() {
 				pthread_cancel(share_updater_thread);
 				share_updater_thread = 0;
 
-				int i;
+				int8_t i;
 				for (i=0;i<CAID_KEY;i++)
 					cc_free_reported_carddata(reported_carddatas[i], NULL, 0);
 		}
