+++ oscam-log.c
@@ -323,7 +323,7 @@ void cs_close_log(void)
 	if (!fp) return;
 
 	//Wait for log close:
-	int i = 2;
+	int8_t i = 2;
 	while (ll_count(log_list) > 0 && i) {
 		cs_sleepms(500);
 		i--;
